+++ R/tabpart_mains.R
@@ -1,8 +1,21 @@
 #' @importFrom officer fp_cell fp_par fp_text fp_border
 complex_tabpart <- function( data, col_keys = names(data),
-                             default_pr_text = fp_text(),
-                             default_pr_par = fp_par(),
-                             default_pr_cell = fp_cell(border = fp_border(color = "transparent")),
+                             default_pr_text = fp_text(
+                               font.family = flextable_global$defaults$font.family,
+                               font.size = flextable_global$defaults$font.size,
+                               color = flextable_global$defaults$font.color
+                             ),
+                             default_pr_par = fp_par(
+                               text.align = flextable_global$defaults$text.align,
+                               padding.left = flextable_global$defaults$padding.left,
+                               padding.right = flextable_global$defaults$padding.right,
+                               padding.bottom = flextable_global$defaults$padding.bottom,
+                               padding.top = flextable_global$defaults$padding.top
+                             ),
+                             default_pr_cell = fp_cell(
+                               background.color = flextable_global$defaults$background.color,
+                               border = fp_border(color = flextable_global$defaults$border.color)
+                             ),
                              cwidth = NULL, cheight = NULL ){
 
   pr_cell_init <- as_struct(nrow(data), col_keys, default_pr_cell, cell_struct)