+++ R/themes.R
@@ -3,26 +3,23 @@
 #' @title Apply vanilla theme
 #' @description Apply theme vanilla to a flextable
 #' @param x a flextable object
-#' @family flextable theme
+#' @family functions related to themes
 #' @examples
 #' ftab <- flextable(iris)
 #' ftab <- theme_vanilla(ftab)
 theme_vanilla <- function(x){
   if( !inherits(x, "flextable") ) stop("theme_vanilla supports only flextable objects.")
-  std_b <- fp_border(width = 1, color = "#333333")
+  std_b <- fp_border(width = 2, color = "#333333")
+  thin_b <- fp_border(width = 0.5, color = "#333333")
   x <- border_remove(x)
 
-  x <- hline( x, border = std_b, part = "all")
+  x <- hline( x, border = thin_b, part = "all")
   x <- hline_top( x, border = std_b, part = "header" )
-  x <- bg(x = x, bg = "transparent", part = "all")
-  x <- color(x = x, color = "#111111", part = "all")
-  x <- fontsize(x = x, size = 11, part = "all")
-  x <- font(x = x, fontname = ifelse( font_family_exists(font_family = "Roboto"), "Roboto", "Arial" ), part = "all")
+  x <- hline_bottom( x, border = std_b, part = "header" )
+  x <- hline_bottom( x, border = std_b, part = "body" )
   x <- bold(x = x, bold = TRUE, part = "header")
   x <- align_text_col(x, align = "left", header = TRUE)
   x <- align_nottext_col(x, align = "right", header = TRUE)
-  x <- padding(x = x, padding.left = 5, padding.right = 5,
-               padding.bottom = 2, padding.top = 2, part = "all")
   x
 }
 
@@ -31,7 +28,7 @@ theme_vanilla <- function(x){
 #' @title Apply box theme
 #' @description Apply theme box to a flextable
 #' @param x a flextable object
-#' @family flextable theme
+#' @family functions related to themes
 #' @examples
 #' ftab <- flextable(iris)
 #' ftab <- theme_box(ftab)
@@ -45,13 +42,10 @@ theme_box <- function(x){
   x <- border_inner_h(x, border = std_border, part="all")
   x <- border_inner_v(x, border = std_border, part="all")
 
-  x <- bg(x = x, bg = "transparent", part = "all")
   x <- bold(x = x, bold = TRUE, part = "header")
   x <- italic(x = x, italic = TRUE, part = "footer")
   x <- align_text_col(x, align = "left", header = TRUE)
   x <- align_nottext_col(x, align = "right", header = TRUE)
-  x <- padding(x = x, padding.left = 5, padding.right = 5,
-               padding.bottom = 2, padding.top = 2, part = "all")
 
   x
 }
@@ -61,12 +55,13 @@ theme_box <- function(x){
 #' @title Apply alafoli theme
 #' @description Apply theme alafoli to a flextable
 #' @param x a flextable object
-#' @family flextable theme
+#' @family functions related to themes
 #' @examples
 #' ftab <- flextable(iris)
 #' ftab <- theme_alafoli(ftab)
 theme_alafoli <- function(x){
-  if( !inherits(x, "flextable") ) stop("theme_alafoli supports only flextable objects.")
+  if( !inherits(x, "flextable") )
+    stop("theme_alafoli supports only flextable objects.")
   x <- border_remove(x)
   x <- bg(x, bg = "transparent", part = "all")
   x <- color(x, color = "#666666", part = "all")
@@ -85,12 +80,12 @@ theme_alafoli <- function(x){
 #' @title Apply Sith Lord Darth Vader
 #' @description Apply Sith Lord Darth Vader theme to a flextable
 #' @param x a flextable object
-#' @param fontsize font size in pixel
-#' @family flextable theme
+#' @param ... unused
+#' @family functions related to themes
 #' @examples
 #' ftab <- flextable(iris)
 #' ftab <- theme_vader(ftab)
-theme_vader <- function(x, fontsize = 11){
+theme_vader <- function(x, ...){
   if( !inherits(x, "flextable") )
     stop("theme_vader supports only flextable objects.")
 
@@ -99,7 +94,6 @@ theme_vader <- function(x, fontsize = 11){
   x <- color(x, color = "#dfdfdf", part = "all")
   x <- bold(x = x, bold = FALSE, part = "all")
   x <- italic(x = x, italic = FALSE, part = "all")
-  x <- padding(x = x, padding = 4, part = "all")
 
   big_border <- fp_border(color = "#ff0000", width = 3)
 
@@ -115,7 +109,6 @@ theme_vader <- function(x, fontsize = 11){
 
   x <- align_text_col(x, align = "left", header = TRUE)
   x <- align_nottext_col(x, align = "right", header = TRUE)
-  x <- fontsize(x, size = fontsize, part = "all")
   x
 }
 
@@ -124,7 +117,7 @@ theme_vader <- function(x, fontsize = 11){
 #' @description Apply theme zebra to a flextable
 #' @param x a flextable object
 #' @param odd_header,odd_body,even_header,even_body odd/even colors for table header and body
-#' @family flextable theme
+#' @family functions related to themes
 #' @examples
 #' ftab <- flextable(iris)
 #' ftab <- theme_zebra(ftab)
@@ -136,7 +129,6 @@ theme_zebra <- function(x, odd_header = "#CFCFCF", odd_body = "#EFEFEF",
   b_nrow <- nrow_part(x, "body")
 
   x <- border_remove(x)
-  x <- padding(x = x, padding = 2, part = "all")
   x <- align(x = x, align = "center", part = "header")
 
   if(h_nrow > 0 ){
@@ -165,8 +157,6 @@ theme_zebra <- function(x, odd_header = "#CFCFCF", odd_body = "#EFEFEF",
   x <- align_text_col(x, align = "left", header = TRUE)
   x <- align_nottext_col(x, align = "right", header = TRUE)
 
-
-
   x
 }
 
@@ -174,20 +164,20 @@ theme_zebra <- function(x, odd_header = "#CFCFCF", odd_body = "#EFEFEF",
 #' @title Apply tron legacy theme
 #' @description Apply theme tron legacy to a flextable
 #' @param x a flextable object
-#' @family flextable theme
+#' @family functions related to themes
 #' @examples
 #' ftab <- flextable(iris)
 #' ftab <- theme_tron_legacy(ftab)
 theme_tron_legacy <- function(x){
 
-  if( !inherits(x, "flextable") ) stop("theme_tron_legacy supports only flextable objects.")
+  if( !inherits(x, "flextable") )
+    stop("theme_tron_legacy supports only flextable objects.")
   h_nrow <- nrow_part(x, "header")
   f_nrow <- nrow_part(x, "footer")
   b_nrow <- nrow_part(x, "body")
 
   x <- border(x = x, border = fp_border(width = 1, color = "#6FC3DF"),
               part = "all")
-  x <- padding(x = x, padding = 2, part = "all")
   x <- align(x = x, align = "right", part = "all")
   x <- bg(x = x, bg = "#0C141F", part = "all")
 
@@ -214,7 +204,7 @@ theme_tron_legacy <- function(x){
 #' @examples
 #' ftab <- flextable(iris)
 #' ftab <- theme_tron(ftab)
-#' @family flextable theme
+#' @family functions related to themes
 theme_tron <- function(x){
 
   if( !inherits(x, "flextable") ) stop("theme_tron supports only flextable objects.")
@@ -224,7 +214,6 @@ theme_tron <- function(x){
 
   x <- border(x = x, border = fp_border(width = 1, color = "#a4cee5"),
               part = "all")
-  x <- padding(x = x, padding = 2, part = "all")
   x <- align(x = x, align = "right", part = "all")
   x <- bg(x = x, bg = "#000000", part = "all")
 
@@ -248,15 +237,18 @@ theme_tron <- function(x){
 #' @title Apply booktabs theme
 #' @description Apply theme booktabs to a flextable
 #' @param x a flextable object
-#' @param fontsize font size in pixel
+#' @param ... unused
 #' @examples
 #' ftab <- flextable(iris)
 #' ftab <- theme_booktabs(ftab)
-#' @family flextable theme
-theme_booktabs <- function(x, fontsize = 11){
-  if( !inherits(x, "flextable") ) stop("theme_booktabs supports only flextable objects.")
+#' @family functions related to themes
+theme_booktabs <- function(x, ...){
+  if( !inherits(x, "flextable") )
+    stop("theme_booktabs supports only flextable objects.")
+
   big_border <- fp_border(width = 2)
   std_border <- fp_border(width = 1)
+
   h_nrow <- nrow_part(x, "header")
   f_nrow <- nrow_part(x, "footer")
   b_nrow <- nrow_part(x, "body")
@@ -269,20 +261,13 @@ theme_booktabs <- function(x, fontsize = 11){
     x <- hline_bottom(x, border = big_border, part = "header")
   }
   if(f_nrow > 0 ){
-    # x <- hline(x, border = std_border, part = "footer")
     x <- hline_bottom(x, border = big_border, part = "footer")
   }
   if(b_nrow > 0 ){
-    # x <- hline(x, border = std_border, part = "body")
     x <- hline_bottom(x, border = big_border, part = "body")
   }
 
-  x <- padding(x = x, padding.left = 5, padding.right = 5,
-               padding.bottom = 2, padding.top = 2, part = "all")
   x <- align_text_col(x, align = "left", header = TRUE)
   x <- align_nottext_col(x, align = "right", header = TRUE)
-  x <- bg(x = x, bg = "transparent", part = "all")
-  x <- fontsize(x, size = fontsize, part = "all")
   x
-
 }